fix: build corrections
