+++ packages/acquirer-frontend/src/hooks/index.ts
@@ -1,4 +1,5 @@
 import useMerchantId from './useMerchantId'
 import useTable from './useTable'
+import { useMerchantColumns } from './useMerchantColumns'
 
-export { useMerchantId, useTable }
+export { useMerchantId, useTable, useMerchantColumns }
