+++ packages/acquirer-frontend/src/pages/RejectedMerchantRecords/RejectedMerchantRecords.tsx
@@ -1,23 +1,16 @@
-import { useMemo, useState } from 'react'
-import { createColumnHelper, type PaginationState } from '@tanstack/react-table'
+import { useState } from 'react'
+import { type PaginationState } from '@tanstack/react-table'
 import {
   Box,
   Heading,
-  HStack,
   SimpleGrid,
   Stack,
-  Text,
   useDisclosure,
 } from '@chakra-ui/react'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
 import { MerchantRegistrationStatus } from 'shared-lib'
 
-import type { MerchantInfo } from '@/types/merchants'
-import {
-  REGISTRATION_STATUS_COLORS,
-  type RegistrationStatus,
-} from '@/constants/registrationStatus'
 import {
   merchantsFilterSchema,
   type MerchantsFilterForm,
@@ -25,7 +18,7 @@ import {
 import { downloadMerchantsBlobAsXlsx } from '@/utils'
 import { useExportMerchants, useRejectedMerchants } from '@/api/hooks/merchants'
 import { useUsers } from '@/api/hooks/users'
-import { useTable } from '@/hooks'
+import { useTable, useMerchantColumns } from '@/hooks'
 import {
   CustomButton,
   DataTable,
@@ -49,123 +42,16 @@ const RejectedMerchantRecords = () => {
     onClose: onInfoModalClose,
   } = useDisclosure()
 
-  const columns = useMemo(() => {
-    const columnHelper = createColumnHelper<MerchantInfo>()
+  const handleViewDetails = (merchantId: number) => {
+    setSelectedMerchantId(merchantId)
+    onInfoModalOpen()
+  }
 
-    return [
-      columnHelper.accessor('no', {
-        cell: info => info.getValue(),
-        header: 'ID',
-      }),
-      columnHelper.accessor('dbaName', {
-        cell: info => info.getValue(),
-        header: 'Doing Business As Name',
-      }),
-      columnHelper.accessor('registeredName', {
-        cell: info => info.getValue(),
-        header: 'Registered Name',
-      }),
-      columnHelper.accessor('payintoAccountId', {
-        cell: info => info.getValue(),
-        header: 'Payinto Account ID',
-      }),
-      columnHelper.accessor('merchantType', {
-        cell: info => info.getValue(),
-        header: 'Merchant Type',
-      }),
-      columnHelper.accessor('town', {
-        cell: info => info.getValue(),
-        header: 'Town',
-      }),
-      columnHelper.accessor('countrySubdivision', {
-        cell: info => info.getValue(),
-        header: 'Country Subdivision',
-      }),
-      columnHelper.accessor('counterDescription', {
-        cell: info => info.getValue(),
-        header: 'Counter Description',
-      }),
-      columnHelper.accessor('registeredDfspName', {
-        cell: info => info.getValue(),
-        header: 'Registered DFSP Name',
-      }),
-      columnHelper.accessor('maker.name', {
-        cell: info => info.getValue(),
-        header: 'Maker Username',
-      }),
-      columnHelper.accessor('registrationStatus', {
-        cell: info => (
-          <HStack justify='center' spacing='1'>
-            <Box
-              as='span'
-              minW='2'
-              w='2'
-              h='2'
-              borderRadius='full'
-              bg={REGISTRATION_STATUS_COLORS[info.getValue() as RegistrationStatus]}
-            />
-
-            <Text>{info.getValue()}</Text>
-          </HStack>
-        ),
-        header: 'Registration Status',
-      }),
-      columnHelper.accessor('gleif_verified_at', {
-        id: 'gleif-validation-status',
-        cell: ({ row }) => {
-          const verificationDate = row.original.gleif_verified_at
-          const status = row.original.registrationStatus as MerchantRegistrationStatus
-          const isValidated =
-            status === MerchantRegistrationStatus.APPROVED ||
-            status === MerchantRegistrationStatus.REVIEW
-
-          if (verificationDate) {
-            const date = new Date(verificationDate)
-            return (
-              <Stack spacing={1}>
-                <Text color='green.600' fontWeight='semibold' fontSize='sm'>
-                  Verified
-                </Text>
-                <Text fontSize='xs' color='gray.600'>
-                  {date.toLocaleDateString()}{' '}
-                  {date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}
-                </Text>
-              </Stack>
-            )
-          } else if (isValidated) {
-            return (
-              <Text color='green.600' fontWeight='semibold' fontSize='sm'>
-                Validated (Legacy)
-              </Text>
-            )
-          } else {
-            return (
-              <Text color='orange.500' fontWeight='semibold' fontSize='sm'>
-                Pending
-              </Text>
-            )
-          }
-        },
-        header: 'Last Verification',
-      }),
-      columnHelper.display({
-        id: 'view-details',
-        cell: ({ row }) => (
-          <CustomButton
-            mt={{ base: '2', lg: '0' }}
-            mr={{ base: '-2', lg: '3' }}
-            onClick={() => {
-              setSelectedMerchantId(row.original.no)
-              onInfoModalOpen()
-            }}
-          >
-            View Details
-          </CustomButton>
-        ),
-        enableSorting: false,
-      }),
-    ]
-  }, [onInfoModalOpen])
+  const columns = useMerchantColumns({
+    onViewDetails: handleViewDetails,
+    includeRegisteredName: true,
+    includeLEI: false,
+  })
 
   const {
     register,
