+++ packages/acquirer-frontend/vite.config.ts
@@ -4,7 +4,7 @@
 import react from '@vitejs/plugin-react-swc'
 import { defineConfig } from 'vite'
 import tsconfigPaths from 'vite-tsconfig-paths'
-import { configDefaults } from 'vitest/dist/config.js'
+import { configDefaults } from 'vitest/config'
 
 // https://vitejs.dev/config/
 export default defineConfig({
